# specs/openapi.yaml
openapi: 3.1.0
info:
  title: FlowProof API
  version: 0.6.0
paths:
...
  /api/v1/operations/alerts/test:
    post:
      tags:
      - operations
      summary: Send Test Alert
      operationId: send_test_alert_api_v1_operations_alerts_test_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
                title: Response Send Test Alert Api V1 Operations Alerts Test Post
      security:
      - BearerAuth: []
      - SessionCookie: []
  /api/v1/auth/login:
    post:
      tags:
      - auth
      summary: Login
      operationId: login_api_v1_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginIn'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Login Api V1 Auth Login Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/logout:
    post:
      tags:
      - auth
      summary: Logout
      operationId: logout_api_v1_auth_logout_post
      responses:
        '204':
          description: Successful Response
      security:
      - BearerAuth: []
      - SessionCookie: []
  /api/v1/auth/me:
    get:
      tags:
      - auth
      summary: Me
      operationId: me_api_v1_auth_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Me Api V1 Auth Me Get
      security:
      - BearerAuth: []
      - SessionCookie: []
  /api/v1/events:
    post:
      tags:
      - events
      summary: Ingest Event
      operationId: ingest_event_api_v1_events_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BusinessEventIn'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventReceipt'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - BearerAuth: []
      - SessionCookie: []
  /api/v1/entities/{entity_type}/{entity_id}/timeline:
    get:
      tags:
      - entities
      summary: Entity Timeline
      operationId: entity_timeline_api_v1_entities__entity_type___entity_id__timeline_get
      parameters:
      - name: entity_type
        in: path
        required: true
        schema:
          type: string
          title: Entity Type
      - name: entity_id
        in: path
        required: true
        schema:
          type: string
          title: Entity Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Entity Timeline Api V1 Entities  Entity Type   Entity
                  Id  Timeline Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - BearerAuth: []
      - SessionCookie: []
  /api/v1/correlations/{correlation_id}/evaluate:
    post:
      tags:
      - evaluation
      summary: Evaluate Correlation
      operationId: evaluate_correlation_api_v1_correlations__correlation_id__evaluate_post
      parameters:
      - name: correlation_id
        in: path
        required: true
        schema:
          type: string
          title: Correlation Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Evaluate Correlation Api V1 Correlations  Correlation
                  Id  Evaluate Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - BearerAuth: []
      - SessionCookie: []
  /api/v1/deadline-jobs:
    get:
      tags:
      - scheduler
      summary: List Deadline Jobs
      operationId: list_deadline_jobs_api_v1_deadline_jobs_get
      parameters:
      - name: state
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: State
      - name: correlation_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Correlation Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response List Deadline Jobs Api V1 Deadline Jobs Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - BearerAuth: []
      - SessionCookie: []
  /api/v1/deadline-jobs/{job_id}:
    get:
      tags:
      - scheduler
      summary: Deadline Job Detail
      operationId: deadline_job_detail_api_v1_deadline_jobs__job_id__get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Deadline Job Detail Api V1 Deadline Jobs  Job Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - BearerAuth: []
      - SessionCookie: []
  /api/v1/incidents:
    get:
      tags:
      - incidents
      summary: List Incidents
      operationId: list_incidents_api_v1_incidents_get
      parameters:
      - name: status_filter
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status Filter
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response List Incidents Api V1 Incidents Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - BearerAuth: []
      - SessionCookie: []
  /api/v1/incidents/{incident_id}:
    get:
      tags:
      - incidents
      summary: Incident Detail
      operationId: incident_detail_api_v1_incidents__incident_id__get
      parameters:
      - name: incident_id
        in: path
        required: true
        schema:
          type: string
          title: Incident Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Incident Detail Api V1 Incidents  Incident Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - BearerAuth: []
      - SessionCookie: []
  /api/v1/blast-radius:
    get:
      tags:
      - incidents
      summary: Blast Radius
      operationId: blast_radius_api_v1_blast_radius_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Blast Radius Api V1 Blast Radius Get
      security:
      - BearerAuth: []
      - SessionCookie: []
  /api/v1/policies:
    get:
      tags:
      - policies
      summary: Active Policies
      operationId: active_policies_api_v1_policies_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Active Policies Api V1 Policies Get
      security:
      - BearerAuth: []
      - SessionCookie: []
...
  /api/v1/demo/false-200/start:
    post:
      tags:
      - demo
      summary: Start False 200 Demo
      operationId: start_false_200_demo_api_v1_demo_false_200_start_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Start False 200 Demo Api V1 Demo False 200 Start Post
      security:
      - BearerAuth: []
      - SessionCookie: []
...
components:
  schemas:
...
    BusinessEventIn:
      properties:
        idempotency_key:
          type: string
          maxLength: 255
          minLength: 1
          title: Idempotency Key
        correlation_id:
          type: string
          maxLength: 255
          minLength: 1
          title: Correlation Id
        entity_type:
          type: string
          pattern: ^[a-z][a-z0-9_.-]{0,63}$
          title: Entity Type
        entity_id:
          type: string
          maxLength: 255
          minLength: 1
          title: Entity Id
        event_type:
          type: string
          pattern: ^[a-z][a-z0-9_.-]{0,127}$
          title: Event Type
        occurred_at:
          type: string
          format: date-time
          title: Occurred At
        source:
          $ref: '#/components/schemas/SourceIn'
        payload:
          additionalProperties: true
          type: object
          maxProperties: 64
          title: Payload
      additionalProperties: false
      type: object
      required:
      - idempotency_key
      - correlation_id
      - entity_type
      - entity_id
      - event_type
      - occurred_at
      - source
      title: BusinessEventIn
...
    EventReceipt:
      properties:
        id:
          type: string
          title: Id
        duplicate:
          type: boolean
          title: Duplicate
        content_hash:
          type: string
          title: Content Hash
      type: object
      required:
      - id
      - duplicate
      - content_hash
      title: EventReceipt
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LoginIn:
      properties:
        name:
          type: string
          maxLength: 128
          minLength: 1
          title: Name
        password:
          type: string
          maxLength: 1024
          minLength: 1
          title: Password
      additionalProperties: false
      type: object
      required:
      - name
      - password
      title: LoginIn
...
    SourceIn:
      properties:
        system:
          type: string
          maxLength: 64
          minLength: 1
          title: System
        workflow_id:
          anyOf:
          - type: string
            maxLength: 255
          - type: 'null'
          title: Workflow Id
        workflow_version:
          anyOf:
          - type: string
            maxLength: 255
          - type: 'null'
          title: Workflow Version
        execution_id:
          anyOf:
          - type: string
            maxLength: 255
          - type: 'null'
          title: Execution Id
        node_name:
          anyOf:
          - type: string
            maxLength: 255
          - type: 'null'
          title: Node Name
      additionalProperties: false
      type: object
      required:
      - system
      title: SourceIn
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API credential issued to a human or service principal.
    SessionCookie:
      type: apiKey
      in: cookie
      name: flowproof_session
      description: Human browser session; state-changing requests also require X-CSRF-Token.